backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["VADIM POLEGAEV"]

[tool.poetry.dependencies]
python = "^3.13"
django = "5.2"
python-dotenv = "1.0.1"
djangorestframework = "3.15.2"
django-filter = "24.3"
djangorestframework-simplejwt = "^5.5.0"
pillow = "^11.2.1"
drf-extra-fields = "^3.7.0"
djoser = "^2.3.1"
gunicorn = "^23.0.0"
psycopg2-binary = "^2.9.10"
psycopg2 = "^2.9.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
